Highlights: - routing - fixed bugs - better dark theme (with automatic system dark mode detection)
