--- blocs todo.txt
@@ -1,20 +1,0 @@
-Done:
-- tiny crown on chat owner
-- "fix" terrible codebase (More specifically Modals)
-- add to chat button on profiles 
-- add users to gc via search
-- Theme modal (for oobe and settings)
-- OOBE
-- image button (simpler way to upload images)
-- BGM
-- BGM + SFX modal
-- basic modal
-- authors for music in bgm
-- disable bgm by default
-
-Might do later on:
-- custom themes
-
-tnix or someone:
-- Fix modal importing
-- fix stupid button inperfections and css
