+++ package.json
@@ -6,12 +6,15 @@
 	"scripts": {
 		"dev": "vite",
 		"dev-host": "vite --host",
-		"build": "vite build",
+		"dev:routify": "routify",
+    	"build": "routify -b && vite build",
 		"preview": "vite preview",
 		"format": "prettier -w ."
 	},
 	"devDependencies": {
+		"@roxi/routify": "^2.18.12",
 		"@sveltejs/vite-plugin-svelte": "^1.4.0",
+		"npm-run-all": "^4.1.5",
 		"prettier": "^2.8.1",
 		"prettier-plugin-svelte": "^2.9.0",
 		"svelte": "^3.49.0",
@@ -19,7 +22,6 @@
 		"vite": "^3.0.0"
 	},
 	"dependencies": {
-		"lucide-svelte": "^0.260.0",
 		"twemoji": "^14.0.2"
 	}
 }
