+++ src/assets/style.css
@@ -57,7 +57,7 @@
 body {
 	margin: 0;
 	overflow: hidden;
-	background-color: #f9a636;
+	background-color: black;
 }
 
 #app,
