+++ src/lib/ProfileView.svelte
@@ -1,12 +1,12 @@
 <script>
 	import loadProfile from "./loadProfile.js";
-	import {profileClicked, mainPage as page, ulist, user} from "./stores.js";
+	import {ulist, user} from "./stores.js";
 	import Loading from "./Loading.svelte";
 	import Container from "./Container.svelte";
 	import PFP from "./PFP.svelte";
 	import {levels} from "./formatting.js";
 	import LiText from "./LiText.svelte";
-	import {tick} from "svelte";
+	import {goto} from '@roxi/routify';
 
 	export let username = "";
 	export let profile = null;
@@ -33,10 +33,7 @@
 					<button
 						class="clickable-pfp"
 						on:click={async () => {
-							page.set("");
-							await tick();
-							profileClicked.set(data._id);
-							page.set("profile");
+							$goto(`/users/${data._id}`);
 						}}
 					>
 						<PFP
