+++ src/lib/Setup.svelte
@@ -1,29 +1,31 @@
 <!-- Boring orange screen with login and signup. -->
 <script>
 	import {
-		screen,
+		setup as setupShown,
 		setupPage as page,
-		mainPage,
 		modalShown,
 		modalPage,
 		authHeader,
 		user,
-	} from "../lib/stores.js";
-	import * as clm from "../lib/clmanager.js";
+	} from "./stores.js";
+	import * as clm from "./clmanager.js";
 	const link = clm.link;
 	// @ts-ignore
 	window.clm = clm;
 
-	import unloadedProfile from "../lib/unloadedprofile.js";
+	import unloadedProfile from "./unloadedprofile.js";
 
 	import meowerLogo from "../assets/logo.svg";
 	import meowy from "../assets/meowy.svg";
 
 	import {tick, onMount} from "svelte";
 	import {fade} from "svelte/transition";
-	import sleep from "../lib/sleep.js";
-	import version from "../lib/version.js";
-	import * as BGM from "../lib/BGM.js";
+	import sleep from "./sleep.js";
+	import version from "./version.js";
+	import * as BGM from "./BGM.js";
+
+	import {isActive, goto} from '@roxi/routify';
+	
 
 	let logo,
 		setup,
@@ -37,8 +39,10 @@
 		await new Promise(resolve => link.once("connected", resolve));
 	}
 
-	let rememberMe = false;
 	let acceptTerms = false;
+
+	let requireLogin = false;
+	$: requireLogin = ($isActive("./setup") || $isActive("./inbox") || $isActive("./chats", {}, { strict: false }));
 
 	onMount(() => {
 		page.subscribe(async value => {
@@ -105,8 +109,13 @@
 		user.set(unloadedProfile());
 		loginStatus = "";
 		page.set("blank");
-		await sleep(600);
-		page.set("welcome");
+		await sleep(500);
+		if (requireLogin || $isActive("./index")) {
+			page.set("welcome");
+		} else {
+			page.set("blank");
+			setupShown.set(false);
+		}
 	}
 
 	/**
@@ -153,23 +162,17 @@
 							token: val.payload.token,
 						});
 
-						if (
-							rememberMe ||
-							localStorage.getItem("meower_savedusername") ===
-								username
-						) {
-							localStorage.setItem(
-								"meower_savedusername",
-								username
-							);
-							localStorage.setItem(
-								"meower_savedpassword",
-								val.payload.token
-							);
-						}
+						localStorage.setItem(
+							"meower_savedusername",
+							username
+						);
+						localStorage.setItem(
+							"meower_savedpassword",
+							val.payload.token
+						);
 
 						BGM.playBGM($user.bgm_song);
-						screen.set("main");
+						setupShown.set(false);
 					} catch (e) {
 						localStorage.clear();
 						console.error(
@@ -256,7 +259,6 @@
 				{#if localStorage.getItem("meower_savedusername")}
 					<button
 						on:click={() => {
-							rememberMe = true;
 							doLogin(
 								localStorage.getItem("meower_savedusername"),
 								localStorage.getItem("meower_savedpassword"),
@@ -270,16 +272,22 @@
 					>
 					<p class="small">{loginStatus}</p>
 				{/if}
-				<button
-					on:click={() => {
-						loginStatus = "";
-						page.set("blank");
-						screen.set("main");
-					}}>Skip</button
-				>
-				<p class="small">
-					(Several features will be unavailable while not logged in.)
-				</p>
+				{#if !requireLogin}
+					<button
+						on:click={() => {
+							loginStatus = "";
+							page.set("blank");
+							setupShown.set(false);
+						}}>Skip</button
+					>
+					<p class="small">
+						(Several features will be unavailable while not logged in.)
+					</p>
+				{:else}
+					<p class="small">
+						(You need to be logged in for this page.)
+					</p>
+				{/if}
 				<div>
 					<p class="small">Meower Svelte v{version}</p>
 					<img src={meowy} alt="" height="64" />
@@ -303,16 +311,7 @@
 				}}
 			>
 				<input type="text" placeholder="Username" maxlength="20" />
-				<br />
 				<input type="password" placeholder="Password" maxlength="255" />
-				<p class="checkboxes">
-					<input
-						id="remember-me"
-						type="checkbox"
-						bind:checked={rememberMe}
-					/>
-					<label for="remember-me"> Save this login </label>
-				</p>
 				<span class="login-status">{loginStatus}</span>
 				<div class="buttons">
 					<button
@@ -381,20 +380,18 @@
 
 								loginStatus = "";
 
-								if (rememberMe) {
-									localStorage.setItem(
-										"meower_savedusername",
-										username
-									);
-									localStorage.setItem(
-										"meower_savedpassword",
-										val.payload.token
-									);
-								}
+								localStorage.setItem(
+									"meower_savedusername",
+									username
+								);
+								localStorage.setItem(
+									"meower_savedpassword",
+									val.payload.token
+								);
 
-								mainPage.set("oobe");
+								$goto("/setup");
 								await sleep(10);
-								screen.set("main");
+								setupShown.set(false);
 							} else {
 								loginStatus = "Unexpected error logging in!";
 							}
@@ -425,7 +422,6 @@
 				}}
 			>
 				<input type="text" placeholder="Username" maxlength="20" />
-				<br />
 				<input
 					type="password"
 					placeholder="Password"
@@ -433,13 +429,6 @@
 					maxlength="255"
 				/>
 				<p class="checkboxes">
-					<input
-						id="remember-me"
-						type="checkbox"
-						bind:checked={rememberMe}
-					/>
-					<label for="remember-me"> Save this login </label>
-					<br />
 					<input
 						id="accept-terms"
 						type="checkbox"
@@ -588,5 +577,15 @@
 	.checkboxes {
 		text-align: left;
 		font-size: 90%;
+	}
+
+	button, input {
+		margin-bottom: 0.2cm;
+	}
+
+	.buttons {
+		display: grid;
+		grid-auto-flow: column;
+		grid-column-gap: 8px;
 	}
 </style>
