+++ src/lib/unloadedprofile.js
@@ -3,11 +3,14 @@
  *
  * @returns {object} The unloaded profile.
  */
+
+import { systemDarkMode } from "./responsiveness.js";
+
 export default function unloadedProfile() {
 	return {
 		name: null,
 		theme: "orange",
-		mode: true,
+		mode: !systemDarkMode(),
 		sfx: true,
 		bgm: false,
 		bgm_song: 2,
