+++ src/pages/@[username].svelte
@@ -1,0 +1,5 @@
+<script>
+    import {params, goto} from '@roxi/routify';
+    
+    $goto(`/users/${$params.username.replace("@", "")}`);
+</script>
