+++ src/pages/about.svelte
@@ -1,0 +1,8 @@
+<script>
+	import { goto } from "@roxi/routify";
+	import {onMount} from "svelte";
+
+    onMount(() => {
+        $goto("/home");
+    })
+</script>
