--- src/screens/settings/Profile.svelte
@@ -1,6 +1,0 @@
-<script>
-	import {user} from "../../lib/stores.js";
-	import ProfileView from "../../lib/ProfileView.svelte";
-</script>
-
-<ProfileView username={$user.name} />
